# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo phone application",
        "version": "0.9"
    },
    "paths": {
        "/api/phone/index": {
            "post": {
                "tags": [
                    "Phone"
                ],
                "summary": "Список всех телефонов с поиском по полям предложений.",
                "description": "После запроса будет получен список всех телефонов, с поиском по полям предложений",
                "operationId": "App\\Http\\Controllers\\API\\PhoneController::index",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "color": {
                                        "description": "Цвет телефона.",
                                        "type": "string",
                                        "default": "black",
                                        "nullable": true
                                    },
                                    "memory": {
                                        "description": "Память телефона.",
                                        "type": "integer",
                                        "default": "32",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Список телефонов.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OAPhoneResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/phone/create": {
            "post": {
                "tags": [
                    "Phone"
                ],
                "summary": "Создание модели телефона.",
                "description": "После запроса будет создан телефон",
                "operationId": "App\\Http\\Controllers\\API\\PhoneController::create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "brand": {
                                        "description": "Бренд",
                                        "type": "string",
                                        "default": "Samsung"
                                    },
                                    "model": {
                                        "description": "Модель",
                                        "type": "string",
                                        "default": "S9"
                                    },
                                    "price": {
                                        "description": "Цена",
                                        "type": "integer",
                                        "default": "45000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JSON c полем 'message', success - все прошло успешно, failed - что-то пошло не так"
                    }
                }
            }
        },
        "/api/phone/delete/{id}": {
            "get": {
                "tags": [
                    "Phone"
                ],
                "summary": "Удаление телефона.",
                "description": "Удаление телефона по его ID",
                "operationId": "App\\Http\\Controllers\\API\\PhoneController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор телефона",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON c полем 'message', success - все прошло успешно, failed - что-то пошло не так"
                    }
                }
            }
        },
        "/api/phoneOffer/create": {
            "post": {
                "tags": [
                    "PhoneOffer"
                ],
                "summary": "Создание предложения телефона.",
                "description": "После запроса будет создано предложение телефона.",
                "operationId": "App\\Http\\Controllers\\API\\PhoneOfferController::create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "memory": {
                                        "description": "Память.",
                                        "type": "integer",
                                        "default": "32"
                                    },
                                    "color": {
                                        "description": "Цвет",
                                        "type": "string",
                                        "default": "black"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JSON c полем 'message', success - все прошло успешно, failed - что-то пошло не так"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OAPhoneResource": {
                "title": "Телефон",
                "properties": {
                    "brand": {
                        "description": "Бренд.",
                        "type": "string",
                        "example": "Samsung"
                    },
                    "model": {
                        "description": "Модель.",
                        "type": "string",
                        "example": "S10"
                    },
                    "price": {
                        "description": "Стоимость.",
                        "type": "integer",
                        "example": "56000.0"
                    },
                    "year": {
                        "description": "Год выпуска.",
                        "type": "integer",
                        "example": "2010"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}